sys/src/9: fix warnings
k10/mp.c:221:3: warning: comparison is always false due to limited range of data type
386/uarti8250.c:779:3: warning: ‘return’ with a value, in function returning void
ip/ipifc.c:1601:1: warning: ‘adddefroute6’ defined but not used
ip/tcp.c:1008:1: warning: ‘tcpflag’ defined but not used
port/cpu_buffer.c:68:14: warning: comparison of unsigned expression >= 0 is always true
port/devmnt.c:136:2: warning: comparison of unsigned expression < 0 is always false
port/devmnt.c:1048:3: warning: comparison is always false due to limited range of data type
port/devproc.c:603:6: warning: comparison is always true due to limited range of data type
port/devsd.c:1447:9: warning: comparison is always true due to limited range of data type
port/devsegment.c:625:6: warning: comparison is always true due to limited range of data type
port/devsrv.c:275:6: warning: comparison is always true due to limited range of data type
port/devuart.c:532:6: warning: comparison is always true due to limited range of data type
port/netif.c:425:11: warning: comparison is always true due to limited range of data type

+++ sys/src/9/port/devproc.c
@@ -600,7 +600,7 @@ procwstat(Chan *c, uint8_t *db, int32_t n)
 		else
 			kstrdup(&p->user, d->uid);
 	}
-	if(d->mode != ~0UL)
+	if(d->mode != (uint32_t)~0UL)
 		p->procmode = d->mode&0777;
 
 	poperror();
